Phani | Rename folder otNotes in components
