+++ micro-frontends/src/next-ui/Components/OtNotes/OtNotes.scss
